fold all updates into one debounce
