+++ services/ui-src/src/components/app/AppRoutes.tsx
@@ -14,6 +14,7 @@ import {
 import { useStore } from "utils";
 import { useEffect } from "react";
 import { useFlags } from "launchdarkly-react-client-sdk";
+import { ReportAutosaveProvider } from "components/report/ReportAutosaveProvider";
 
 export const AppRoutes = () => {
   const { userIsAdmin } = useStore().user ?? {};
@@ -30,33 +31,37 @@ export const AppRoutes = () => {
   return (
     <main id="main-content" tabIndex={-1}>
       <AdminBannerProvider>
-        <Routes>
-          {/* General Routes */}
-          <Route path="/" element={<HomePage />} />
-          <Route
-            path="/admin"
-            element={!userIsAdmin ? <Navigate to="/profile" /> : <AdminPage />}
-          />
-          <Route path="/profile" element={<ProfilePage />} />
-          <Route path="/help" element={<HelpPage />} />
-          <Route path="*" element={<NotFoundPage />} />
-          <Route
-            path="/report/:reportType/:state"
-            element={<DashboardPage />}
-          />
-          <Route path="/report/QMS" element={<CreateReportOptions />} />
-          {isPdfActive && (
+        <ReportAutosaveProvider>
+          <Routes>
+            {/* General Routes */}
+            <Route path="/" element={<HomePage />} />
             <Route
-              path="/report/:reportType/:state/:reportId/export"
-              element={<ExportedReportPage />}
+              path="/admin"
+              element={
+                !userIsAdmin ? <Navigate to="/profile" /> : <AdminPage />
+              }
             />
-          )}
-          <Route
-            path="/report/:reportType/:state/:reportId/:pageId?"
-            element={<ReportPageWrapper />}
-          />
-          {/* TO DO: Load pageId by default? */}
-        </Routes>
+            <Route path="/profile" element={<ProfilePage />} />
+            <Route path="/help" element={<HelpPage />} />
+            <Route path="*" element={<NotFoundPage />} />
+            <Route
+              path="/report/:reportType/:state"
+              element={<DashboardPage />}
+            />
+            <Route path="/report/QMS" element={<CreateReportOptions />} />
+            {isPdfActive && (
+              <Route
+                path="/report/:reportType/:state/:reportId/export"
+                element={<ExportedReportPage />}
+              />
+            )}
+            <Route
+              path="/report/:reportType/:state/:reportId/:pageId?"
+              element={<ReportPageWrapper />}
+            />
+            {/* TO DO: Load pageId by default? */}
+          </Routes>
+        </ReportAutosaveProvider>
       </AdminBannerProvider>
     </main>
   );
